auto_cibus_web_interface/requirements.txt
flask==3.0.3
python-crontab
